fix: allow the code to build when non-default features are turned off
