+++ src/bundle/mod.rs
@@ -16,6 +16,7 @@
 
 pub use sigstore_protobuf_specs::dev::sigstore::bundle::v1::Bundle;
 
+pub(crate) mod intoto;
 mod models;
 
 #[cfg_attr(docsrs, doc(cfg(feature = "sign")))]
